upgrade: tiptap v2
*Update dependencies and enable using tiptap v2*

* uninstall tiptap v1 install v2
* replace the npm imports

*Migrate building blocks*

* port marks to the new syntax
  * links with our custom link behavior
* port tiptap1 collaboration
  We still want to use our session and sync mechanism.

*Port the menubar*

* specify args for isActive function directly rather than setting a function.
* use the new commands.
* make use of the editor instance inside the MenuBar component.
* use the editor rather than slots for command, focused etc.

*Use the new extensions of tiptap2*

* blockquotes
* code block

*Port our own extensions*

* User Colors
* Emoji Picker
* Keymap
* Task lists

*Convert prosemirror names*

Now the tiptap node name for a number of elements
does not match the prosemirror one.
Rename the prosemirror marks and nodes to camel case
rather than renaming every camel case tiptap command.

*Upgrade image node and ImageView*

Quite a bit of the syntax changed.
We now need a wrapping   `<node-view-wrapper>` element.
Pretty good docs at https://tiptap.dev/guide/node-views/vue#render-a-vue-component
We also need to handle the async action.
It will run the action on it's own.
So in `clickIcon()` we need to test if the action returned anything.