fix: Todo list toggle
Bring the todo list back to life.

It is behaving somewhat strangely when mixed with normal lists.
Not sure if this is a regression or not.

Also introduced a cypress test that tests the editor
without rendering the page and logging in and so on.
It is very fast and fairly easy to write.